+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.1.10 - 2025-10-29
+
+### Added
+
+- add a new `VrchAudioWebSocketChannelLoaderNode` node (display name: `AUDIO WebSocket Channel Loader @ vrch.ai`)
+
+### Updated
+
+- update `docs/websocket_nodes.md` to document the new audio WebSocket channel loader node
+
 ## 1.1.9 - 2025-10-23
 
 ### Updated